Update Marathon3-java And xml files
In the testcase2 and testcase3. The dependsOnMethod entered as just the fileName. This gave the method not detected error in TestNG (TestNG exception). Solution is to add the complete packageName.className.methodName. 

PS: if the methods are in the same class then, the dependsOnMethods doesn't have to be mentioned with the packageName.className. In our case, these classes are different but comes under same package(marathon3). So, giving the methodName with package Name solved the issue.